refactor notification service

+++ Lotus/Microservices/notification.service/src/services/RabbitMQ/connectRabbitMQ.js
@@ -1,10 +1,17 @@
 const amqp = require('amqplib/callback_api');
-const { USER: User } = require('./../../database/models/notification');
-const {NOTIFICATION} = require("../../database/Models/notification");
-const {Mailer} = require("../mailer/mailer");
+const { NOTIFICATION } = require("../../database/Models/notification");
+const { Mailer } = require("../mailer/mailer");
 
-const RABBITMQ_HOST = process.env.RABBITMQ_HOST == null ? "localhost" : process.env.RABBITMQ_HOST;
-const RABBITMQ_PORT = process.env.RABBITMQ_PORT == null ? 5672 : process.env.RABBITMQ_PORT;
+const RABBITMQ_HOST = process.env.RABBITMQ_HOST || "localhost";
+const RABBITMQ_PORT = process.env.RABBITMQ_PORT || 5672;
+
+const QUEUES = {
+    NotifyUserRegisteredQueue: 'NotifyUserRegisteredQueue',
+    UserNotificationQueue: 'UserNotificationQueue',
+    SystemNotificationQueue: 'SystemNotificationQueue',
+    EmailNotificationQueue: 'EmailNotificationQueue',
+    LastEmailNotificationQueue: 'LastEmailNotificationQueue'
+};
 
 function connectRabbitMQ() {
     amqp.connect(`amqp://${RABBITMQ_HOST}:${RABBITMQ_PORT}`, function(error0, connection) {
@@ -16,34 +23,12 @@ function connectRabbitMQ() {
                 throw error1;
             }
 
-            const NotifyUserRegisteredQueue = 'NotifyUserRegisteredQueue';
-            const UserNotificationQueue = 'UserNotificationQueue';
-            const SystemNotificationQueue = 'SystemNotificationQueue';
-            const EmailNotificationQueue = 'EmailNotificationQueue';
-            const LastEmailNotificationQueue = 'LastEmailNotificationQueue';
-
-            channel.assertQueue(UserNotificationQueue, {
-                durable: false
-            });
-            channel.assertQueue(SystemNotificationQueue, {
-                durable: false
-            });
-            channel.assertQueue(EmailNotificationQueue, {
-                durable: false
-            });
-            channel.assertQueue(NotifyUserRegisteredQueue, {
-                durable: false
-            });
-            channel.assertQueue(LastEmailNotificationQueue, {
-                durable: false
+            Object.values(QUEUES).forEach(queue => {
+                channel.assertQueue(queue, { durable: false });
             });
 
-
-            channel.consume(NotifyUserRegisteredQueue, async function (msg) {
-                const data = JSON.parse(msg.content.toString());
-                const user_id = data.USER_ID;
-                const email = data.EMAIL;
-                const username = data.USERNAME;
+            channel.consume(QUEUES.NotifyUserRegisteredQueue, async function (msg) {
+                const { USER_ID: user_id, EMAIL: email, USERNAME: username } = JSON.parse(msg.content.toString());
                 Mailer.sendEmailUserRegistered(email, username);
                 await NOTIFICATION.create({
                     AUTHOR: "System notification",
@@ -51,28 +36,15 @@ function connectRabbitMQ() {
                     CONTENT: "Welcome to Lotus! Hi! We are glad to welcome you to Lotus. You have successfully registered. Start enjoying all the benefits of Lotus right now!",
                     IMAGE: "system_notification.png",
                 });
-            }, {
-                noAck: true
-            });
+            }, { noAck: true });
 
-            channel.consume(UserNotificationQueue, async function (msg) {
-                const data = JSON.parse(msg.content.toString());
-                await NOTIFICATION.create({
-                    AUTHOR: data.AUTHOR,
-                    USER_ID: data.USER_ID,
-                    CONTENT: data.CONTENT,
-                    IMAGE: data.IMAGE,
-                });
-            }, {
-                noAck: true
-            });
+            channel.consume(QUEUES.UserNotificationQueue, async function (msg) {
+                const { AUTHOR, USER_ID, CONTENT, IMAGE } = JSON.parse(msg.content.toString());
+                await NOTIFICATION.create({ AUTHOR, USER_ID, CONTENT, IMAGE });
+            }, { noAck: true });
 
-            channel.consume(SystemNotificationQueue, async function (msg) {
-                const data = JSON.parse(msg.content.toString());
-                const email = data.EMAIL;
-                const user_id = data.USER_ID;
-                const username = data.USERNAME;
-                const message = data.MESSAGE;
+            channel.consume(QUEUES.SystemNotificationQueue, async function (msg) {
+                const { EMAIL: email, USER_ID: user_id, USERNAME: username, MESSAGE: message } = JSON.parse(msg.content.toString());
                 Mailer.sendEmailSystemNotification(email, username, message);
                 await NOTIFICATION.create({
                     AUTHOR: "System notification",
@@ -80,38 +52,20 @@ function connectRabbitMQ() {
                     USER_ID: user_id,
                     IMAGE: "system_notification.png",
                 });
-            }, {
-                noAck: true
-            });
+            }, { noAck: true });
 
-            channel.consume(EmailNotificationQueue, async function (msg) {
-                const data = JSON.parse(msg.content.toString());
-                const email = data.EMAIL;
-                const username = data.USERNAME;
-                const message = data.MESSAGE;
+            channel.consume(QUEUES.EmailNotificationQueue, async function (msg) {
+                const { EMAIL: email, USERNAME: username, MESSAGE: message } = JSON.parse(msg.content.toString());
                 Mailer.sendEmailMessage(email, username, message);
-            }, {
-                noAck: true
-            });
+            }, { noAck: true });
 
-            channel.consume(LastEmailNotificationQueue, async function (msg) {
-                const data = JSON.parse(msg.content.toString());
-                const email = data.EMAIL;
-                const user_id = data.USER_ID;
-                const username = data.USERNAME;
-                const message = data.MESSAGE;
+            channel.consume(QUEUES.LastEmailNotificationQueue, async function (msg) {
+                const { EMAIL: email, USER_ID: user_id, USERNAME: username, MESSAGE: message } = JSON.parse(msg.content.toString());
                 Mailer.sendEmailMessage(email, username, message);
-                await NOTIFICATION.destroy({
-                    where: {
-                        USER_ID: user_id
-                    }
-                });
-            }, {
-                noAck: true
-            });
+                await NOTIFICATION.destroy({ where: { USER_ID: user_id } });
+            }, { noAck: true });
         });
     });
 }
-
 
 module.exports = connectRabbitMQ;
